+++ boofuzz/blocks/checksum.py
@@ -1,8 +1,16 @@
 import hashlib
 import struct
+import warnings
 import zlib
 from functools import wraps
 
+try:
+    import crc32c
+except ImportError:
+    # Import guard for systems without crc32c support.
+    warnings.warn("Importing crc32c package failed. Using crc32c checksums will fail.", UserWarning, stacklevel=2)
+    crc32c = None
+    pass
 import six
 
 from .. import exception, helpers, primitives
@@ -35,7 +43,8 @@ class Checksum(primitives.BasePrimitive):
     Args:
         block_name (str): Name of target block for checksum calculations.
         request (s_request): Request this block belongs to.
-        algorithm (Union[str, function], optional): Checksum algorithm to use. (crc32, adler32, md5, sha1, ipv4, udp)
+        algorithm (Union[str, function], optional): Checksum algorithm to use.
+            (crc32, crc32c, adler32, md5, sha1, ipv4, udp)
         length (int, optional): Length of checksum, auto-calculated by default.
             Must be specified manually when using custom algorithm.
         endian (str, optional): Endianness of the bit field (LITTLE_ENDIAN: <, BIG_ENDIAN: >).
@@ -46,7 +55,7 @@ class Checksum(primitives.BasePrimitive):
         ipv4_dst_block_name (str): Required for 'udp' algorithm. Name of block yielding IPv4 destination address.
     """
 
-    checksum_lengths = {"crc32": 4, "adler32": 4, "md5": 16, "sha1": 20, "ipv4": 2, "udp": 2}
+    checksum_lengths = {"crc32": 4, "crc32c": 4, "adler32": 4, "md5": 16, "sha1": 20, "ipv4": 2, "udp": 2}
 
     def __init__(
         self,
@@ -142,6 +151,9 @@ class Checksum(primitives.BasePrimitive):
         if isinstance(self._algorithm, six.string_types):
             if self._algorithm == "crc32":
                 check = struct.pack(self._endian + "L", (zlib.crc32(data) & 0xFFFFFFFF))
+
+            elif self._algorithm == "crc32c":
+                check = struct.pack(self._endian + "L", crc32c.crc32(data))
 
             elif self._algorithm == "adler32":
                 check = struct.pack(self._endian + "L", (zlib.adler32(data) & 0xFFFFFFFF))