+++ CHANGELOG.rst
@@ -10,6 +10,7 @@ Features
 - new primitive `s_bytes` which fuzzes an arbitrary length binary value (similiar to `s_string`)
 - We are now using `Black` for code style standardization
 - Added support for Python 3.8
+- added crc32c as checksum algorithm (Castagnoli)
 
 Fixes
 ^^^^^
@@ -26,6 +27,8 @@ Fixes
 - Fixed setting socket timeout options on Windows
 - If all sockets are exhausted, repeatedly try fuzzing for 4 minutes before failing
 - Fixed CSV logger send and receive data decoding
+- Handle SSL-related exception. Added `ignore_connection_ssl_errors` session attribute that can
+  be set to True to ignore SSL-related error on a test case.
 - Fixed `s_from_file` decoding in Python 2 (the encoding parameter is now depreciated)
 
 v0.1.5